requirements/development.txt
# Develop matching the development guidelines
# -------------------------------------------

black

flake8-eradicate>=1.0
flake8-builtins>=1.5
flake8-isort>=6
flake8-qgis>=1

pre-commit>=3.7,<5

qgis-plugin-ci==2.8.*
